refactor: 디렉토리명 형식 통일
- 복수 형태로 통일 ex) controllers, services
